+++ app/scripts/HorizontalGeneAnnotationsTrack.js
@@ -67,7 +67,6 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
     // create texts
     tile.texts = {};
 
-    // console.log('tile:', tile.tileId, tile.tileData);
     tile.rectGraphics = new PIXI.Graphics();
     tile.textBgGraphics = new PIXI.Graphics();
     tile.textGraphics = new PIXI.Graphics();
@@ -242,7 +241,6 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
         let yMiddle = this.dimensions[1] / 2;
         const geneId = this.geneId(geneInfo);
 
-        // console.log('geneInfo:', geneInfo);
         if (geneInfo[5] === '+') {
           // genes on the + strand drawn above and in a user-specified color or the
           // default blue
@@ -499,8 +497,6 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
 
           const text = tile.texts[geneId];
 
-          // console.log('visible geneName:', geneName);
-
           if (!text) return;
 
           const chrOffset = +td.chrOffset;
@@ -534,9 +530,6 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
             // dimensions are measured for the first tile and not for the second
             const textWidth = text.getBounds().width;
             const textHeight = text.getBounds().height;
-
-            // console.log('textWidth:', textWidth);
-            // console.log('textHeight:', textHeight);
 
             tile.textHeights[geneId] = textHeight;
             tile.textWidths[geneId] = textWidth;